requirements.txt
# Core dependencies
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0

# LangChain ecosystem
langchain-core>=0.1.33
langchain-community>=0.0.29
langchain-huggingface>=0.0.8
langchain-text-splitters>=0.0.1
langchain_chroma >=0.0.2

# PDF processing
pymupdf>=1.23.6
pypdf>=3.17.0


# Database & ORM
sqlalchemy>=2.0.25

# Environment management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.6.1

# Async HTTP client
httpx>=0.26.0

# Optional but recommended
python-multipart>=0.0.6  # For file uploads
typing-extensions>=4.9.0  # Type hints support